pyproject.toml
[project]
name = "pythond"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "discord>=2.3.2",
    "html5lib>=1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
